--- Bullethell/Assets/CameraAspect.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5273cdb9fdc049946b22cabd49775b38
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
