--- Bullethell/Assets/CameraSquareAspect.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f6ee0a0d53f9f294ab84c7d4cf0f9742
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
