+++ Bullethell/Assets/Scripts/CameraBounds.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1a6ffe23a08272843b8a04f2975dcfd1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
