+++ Bullethell/Assets/Sounds.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4cd44a8a934d6904a847b4fc70d3b956
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
